Update Cat Base Stats to Letter Grades
The base stats for each cat are recorded as letter grades and will be
part of how UserCat stats will be derived.
The benefits of this system are that it's easier to pick grades for
stats than it is to pick numbers. And also we can adjust the weights
on grades to numeric value while keeping the relative power levels the
same.